add C-ABI-safe scene marshaling for CUDA

+++ src/geometry/interval.rs
@@ -1,36 +1,36 @@
 //static empty: Interval = Interval::default();
-//static universe: Interval = Interval::new(f64::NEG_INFINITY, f64::INFINITY);
+//static universe: Interval = Interval::new(f32::NEG_INFINITY, f32::INFINITY);
 
 pub struct Interval {
-    pub min: f64,
-    pub max: f64,
+    pub min: f32,
+    pub max: f32,
 }
 
 impl Interval {
     pub fn default() -> Self {
         Self {
-            min: f64::INFINITY,
-            max: f64::NEG_INFINITY,
+            min: f32::INFINITY,
+            max: f32::NEG_INFINITY,
         }
     }
 
-    pub fn new(min: f64, max: f64) -> Self {
+    pub fn new(min: f32, max: f32) -> Self {
         Self { min, max }
     }
 
-    pub fn size(&self) -> f64 {
+    pub fn size(&self) -> f32 {
         self.max - self.min
     }
 
-    pub fn contains(&self, x: f64) -> bool {
+    pub fn contains(&self, x: f32) -> bool {
         self.min <= x && x <= self.max
     }
 
-    pub fn surrounds(&self, x: f64) -> bool {
+    pub fn surrounds(&self, x: f32) -> bool {
         self.min < x && x < self.max    
     }
 
-    pub fn clamp(&self, x: f64) -> f64 {
+    pub fn clamp(&self, x: f32) -> f32 {
         if x < self.min {
             return self.min;
         }
