+++ src/geometry/vec3.rs
@@ -7,17 +7,17 @@ pub type Color = Vec3;
 
 #[derive(Debug, Copy, Clone, PartialEq)]
 pub struct Vec3 {
-    pub e: [f64; 3],
+    pub e: [f32; 3],
 }
 
 impl Vec3 {
     pub fn default() -> Self {
         Self {
-            e: [0_f64, 0_f64, 0_f64]
+            e: [0_f32, 0_f32, 0_f32]
         }
     }
 
-    pub fn new(e0: f64, e1: f64, e2:f64) -> Self {
+    pub fn new(e0: f32, e1: f32, e2:f32) -> Self {
         Self {
             e: [e0, e1, e2]
         }
@@ -33,7 +33,7 @@ impl Vec3 {
         }
     }
 
-    pub fn random_range(min: f64, max: f64) -> Self {
+    pub fn random_range(min: f32, max: f32) -> Self {
         Self {
             e: [
                 random_double_range(min, max),
@@ -45,8 +45,8 @@ impl Vec3 {
 
     fn random_in_unit_sphere() -> Vec3 {
         loop {
-            let p = Self::random_range(-1_f64, 1_f64);
-            if p.length_squared() < 1_f64 { 
+            let p = Self::random_range(-1_f32, 1_f32);
+            if p.length_squared() < 1_f32 { 
                 return p; 
             }
         }
@@ -54,19 +54,19 @@ impl Vec3 {
 
     pub fn random_on_hemisphere(normal: &Vec3) -> Self {
         let on_unit_sphere: Vec3 = Self::random_unit_vector();
-        if Self::dot(&on_unit_sphere, normal) > 0_f64 {
+        if Self::dot(&on_unit_sphere, normal) > 0_f32 {
             return on_unit_sphere;
         } else {
-            return -1_f64 * on_unit_sphere;
+            return -1_f32 * on_unit_sphere;
         }
     }
 
     pub fn reflect(v: &Vec3, n: &Vec3) -> Vec3 {
-        *v - 2_f64 * Self::dot(v, n) * *n
+        *v - 2_f32 * Self::dot(v, n) * *n
     }
 
-    pub fn refract(uv: &Vec3, normal: &Vec3, eta_over_eta_prime: f64) -> Vec3 {
-        let cos_theta: f64 = Self::dot(&-uv, normal).min(1_f64);
+    pub fn refract(uv: &Vec3, normal: &Vec3, eta_over_eta_prime: f32) -> Vec3 {
+        let cos_theta: f32 = Self::dot(&-uv, normal).min(1_f32);
         let r_out_perp: Vec3 = eta_over_eta_prime * (*uv + cos_theta * *normal);
 
         let len_sq = r_out_perp.length_squared();
@@ -75,7 +75,7 @@ impl Vec3 {
             -under_sqrt.sqrt() * *normal
         } else {
             // total internal reflection or numerical error: fallback to zero vector
-            // this is to address potential NaN return from f64::sqrt()
+            // this is to address potential NaN return from f32::sqrt()
             Vec3::default()
         };
 
@@ -86,34 +86,34 @@ impl Vec3 {
         Self::unit_vector(Self::random_in_unit_sphere())
     }
 
-    pub fn x(&self) -> f64 {
+    pub fn x(&self) -> f32 {
         self.e[0]
     }
 
-    pub fn y(&self) -> f64 {
+    pub fn y(&self) -> f32 {
         self.e[1]
     }
 
-    pub fn z(&self) -> f64 {
+    pub fn z(&self) -> f32 {
         self.e[2]
     }
 
-    pub fn length(&self) -> f64 {
+    pub fn length(&self) -> f32 {
         self.length_squared().sqrt()
     }
 
-    pub fn length_squared(&self) -> f64 {
+    pub fn length_squared(&self) -> f32 {
         self.e[0] * self.e[0] + 
         self.e[1] * self.e[1] + 
         self.e[2] * self.e[2]
     } 
 
     pub fn near_zero(&self) -> bool {
-        let s: f64 = 1e-8;
-        (f64::abs(self.e[0]) < s) && (f64::abs(self.e[1]) < s) && (f64::abs(self.e[2]) < s)
+        let s: f32 = 1e-8;
+        (f32::abs(self.e[0]) < s) && (f32::abs(self.e[1]) < s) && (f32::abs(self.e[2]) < s)
     }
 
-    pub fn dot(lhs: &Vec3, rhs: &Vec3) -> f64 {
+    pub fn dot(lhs: &Vec3, rhs: &Vec3) -> f32 {
         lhs.e[0] * rhs.e[0] +
         lhs.e[1] * rhs.e[1] +
         lhs.e[2] * rhs.e[2]
@@ -128,18 +128,18 @@ impl Vec3 {
     }
 
     pub fn unit_vector(v: Vec3) -> Vec3 {
-        (1_f64 / v.length()) * v
+        (1_f32 / v.length()) * v
     }
 
     pub fn random_in_unit_disk() -> Vec3 {
         loop {
             let p: Vec3 = Vec3::new(
-                random_double_range(-1_f64, 1_f64), 
-                random_double_range(-1_f64, 1_f64), 
-                0_f64
+                random_double_range(-1_f32, 1_f32), 
+                random_double_range(-1_f32, 1_f32), 
+                0_f32
             );
 
-            if p.length_squared() < 1_f64 { return p; }
+            if p.length_squared() < 1_f32 { return p; }
         }
     }
 }
@@ -161,7 +161,7 @@ impl ops::Neg for Vec3 {
 }
 
 impl ops::Index<usize> for  Vec3 {
-    type Output = f64;
+    type Output = f32;
 
     fn index(&self, index: usize) -> &Self::Output {
         &self.e[index]
@@ -230,10 +230,10 @@ impl ops::Mul<Vec3> for Vec3 {
     }
 }
 
-impl ops::Mul<f64> for Vec3 {
+impl ops::Mul<f32> for Vec3 {
     type Output = Vec3;
 
-    fn mul(self, rhs: f64) -> Self::Output {
+    fn mul(self, rhs: f32) -> Self::Output {
         Vec3::new(
             self.e[0] * rhs, 
             self.e[1] * rhs, 
@@ -242,7 +242,7 @@ impl ops::Mul<f64> for Vec3 {
     }
 }
 
-impl ops::Mul<Vec3> for f64 {
+impl ops::Mul<Vec3> for f32 {
     type Output = Vec3;
 
     fn mul(self, rhs: Vec3) -> Vec3 {
@@ -264,8 +264,8 @@ impl ops::MulAssign<Vec3> for Vec3 {
     }
 }
 
-impl ops::MulAssign<f64> for Vec3 {
-    fn mul_assign(&mut self, rhs: f64) {
+impl ops::MulAssign<f32> for Vec3 {
+    fn mul_assign(&mut self, rhs: f32) {
         *self = Vec3::new(
             self.e[0] * rhs, 
             self.e[1] * rhs, 
@@ -274,10 +274,10 @@ impl ops::MulAssign<f64> for Vec3 {
     }
 }
 
-impl ops::Div<f64> for Vec3 {
+impl ops::Div<f32> for Vec3 {
     type Output = Vec3;
 
-    fn div(self, rhs: f64) -> Self::Output {
-        (1_f64 / rhs) * self
+    fn div(self, rhs: f32) -> Self::Output {
+        (1_f32 / rhs) * self
     }
 }
