+++ src/hittables/cpu/sphere.rs
@@ -2,17 +2,17 @@ use crate::geometry::vec3::{Vec3, Point3};
 use crate::geometry::ray::Ray;
 use crate::geometry::interval::Interval;
 
-use crate::hittables::hittable::{Hittable, HitRecord};
+use crate::hittables::cpu::hittable::{Hittable, HitRecord};
 use crate::material::material::Material;
 
 pub struct Sphere {
     pub center: Point3,
-    pub radius: f64,
+    pub radius: f32,
     pub mat: Box<dyn Material>
 }
 
 impl Sphere {
-    pub fn new(center: Point3, radius: f64, mat: Box<dyn Material>) -> Self {
+    pub fn new(center: Point3, radius: f32, mat: Box<dyn Material>) -> Self {
         Self { center, radius, mat }
     }
 }
@@ -26,16 +26,16 @@ impl Hittable for Sphere {
     ) -> bool {
         let oc: Vec3 = self.center - r.origin;
 
-        let a: f64 = r.direction.length_squared();
-        let h: f64 = Vec3::dot(&r.direction, &oc);
-        let c: f64 = oc.length_squared() - (self.radius * self.radius);
+        let a: f32 = r.direction.length_squared();
+        let h: f32 = Vec3::dot(&r.direction, &oc);
+        let c: f32 = oc.length_squared() - (self.radius * self.radius);
 
-        let discriminant: f64 = (h * h) - (a * c);
-        if discriminant < 0_f64 { return false; }
-        let sqrtd: f64 = f64::sqrt(discriminant);
+        let discriminant: f32 = (h * h) - (a * c);
+        if discriminant < 0_f32 { return false; }
+        let sqrtd: f32 = f32::sqrt(discriminant);
 
         // Find the nearest root that lies in the acceptable range.
-        let mut root: f64 = (h - sqrtd) / a;
+        let mut root: f32 = (h - sqrtd) / a;
         if !ray_t.surrounds(root) {
             root = (h + sqrtd) / a;
             if !ray_t.surrounds(root) {
